fix: UserService interface 제거
